rename class-encoding => css-class-encoding
