+++ src/css-class-encoding.js
